+++ src/safestclaw/channels/__init__.py
@@ -2,5 +2,6 @@
 
 from safestclaw.channels.base import BaseChannel
 from safestclaw.channels.cli import CLIChannel
+from safestclaw.channels.web import WebChannel
 
-__all__ = ["BaseChannel", "CLIChannel"]
+__all__ = ["BaseChannel", "CLIChannel", "WebChannel"]